Remove `RegulatoryBodyController`
We now move the concerns of `RegulatoryBodyController` into `TopLevelInformationController`. We now have it so that all editing of `Profession` (rather than `ProfessionVersion`) is contained within the top-level information page
