--- src/professions/admin/dto/regulatory-body.dto.ts
@@ -1,12 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class RegulatoryBodyDto {
-  @IsNotEmpty({
-    message: 'professions.form.errors.regulatoryBody.empty',
-  })
-  regulatoryBody: string;
-
-  additionalRegulatoryBody: string;
-
-  change: boolean;
-}
