+++ src/professions/admin/dto/top-level-details.dto.ts
@@ -4,5 +4,12 @@ export class TopLevelDetailsDto {
   @IsNotEmpty({ message: 'professions.form.errors.name.empty' })
   name: string;
 
+  @IsNotEmpty({
+    message: 'professions.form.errors.regulatoryBody.empty',
+  })
+  regulatoryBody: string;
+
+  additionalRegulatoryBody: string;
+
   change: string;
 }
