--- src/professions/admin/interfaces/regulatory-body.template.ts
@@ -1,9 +1,0 @@
-import { SelectItemArgs } from '../../../common/interfaces/select-item-args.interface';
-
-export interface RegulatoryBodyTemplate {
-  regulatedAuthoritiesSelectArgs: SelectItemArgs[];
-  additionalRegulatedAuthoritiesSelectArgs: SelectItemArgs[];
-  change: boolean;
-  captionText: string;
-  errors: object | undefined;
-}
