+++ src/professions/admin/interfaces/top-level-details.template.ts
@@ -1,5 +1,9 @@
+import { SelectItemArgs } from '../../../common/interfaces/select-item-args.interface';
+
 export interface TopLevelDetailsTemplate {
   name: string | null;
+  regulatedAuthoritiesSelectArgs: SelectItemArgs[];
+  additionalRegulatedAuthoritiesSelectArgs: SelectItemArgs[];
   captionText: string;
   change: boolean;
   errors: object | undefined;
