+++ src/professions/admin/registration.controller.ts
@@ -49,7 +49,7 @@ export class RegistrationController {
   @BackLink((request: Request) =>
     request.query.change === 'true'
       ? '/admin/professions/:professionId/versions/:versionId/check-your-answers'
-      : '/admin/professions/:professionId/versions/:versionId/regulatory-body/edit',
+      : '/admin/professions/:professionId/versions/:versionId/scope/edit',
   )
   async edit(
     @Res() res: Response,
@@ -73,7 +73,7 @@ export class RegistrationController {
   @BackLink((request: Request) =>
     request.body.change === 'true'
       ? '/admin/professions/:professionId/versions/:versionId/check-your-answers'
-      : '/admin/professions/:professionId/versions/:versionId/regulatory-body/edit',
+      : '/admin/professions/:professionId/versions/:versionId/scope/edit',
   )
   async update(
     @Res() res: Response,
