--- views/admin/professions/regulatory-body.njk
@@ -1,61 +1,0 @@
-{% extends "admin/base.njk" %}
-
-{% from "govuk/components/input/macro.njk" import govukInput %}
-{% from "govuk/components/select/macro.njk" import govukSelect %}
-{% from "govuk/components/radios/macro.njk" import govukRadios %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-
-{% set bodyClasses = "rpr-internal__page" %}
-
-{% block content %}
-  {% include "../../shared/_errors.njk" %}
-
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-
-      <span class="govuk-caption-l">{{ captionText | t }}</span>
-      <h1 class="govuk-heading-l">{{ ("professions.form.headings.regulatoryBody" | t) }}</h1>
-
-      <form method="post" action="./">
-        <input type="hidden" name="change" value="{{ change }}" />
-
-        {{
-          govukSelect({
-            label: {
-              text: ("professions.form.label.regulatoryBody.regulatedAuthority" | t),
-              classes: "govuk-label--m",
-              isPageHeading: false
-            },
-            id: "regulatoryBody",
-            name: "regulatoryBody",
-            items: regulatedAuthoritiesSelectArgs,
-            errorMessage: errors.regulatoryBody | tError
-          })
-        }}
-
-        {{
-          govukSelect({
-            label: {
-              text: ("professions.form.label.regulatoryBody.additionalAuthority" | t),
-              classes: "govuk-label--m",
-              isPageHeading: false
-            },
-            id: "additionalRegulatoryBody",
-            name: "additionalRegulatoryBody",
-            items: additionalRegulatedAuthoritiesSelectArgs
-          })
-        }}
-
-        {{
-          govukButton({
-            id: "submit-button",
-            type: "Submit",
-            text: ("app.continue" | t)
-          })
-        }}
-      </form>
-
-    </div>
-  </div>
-
-{% endblock %}
