+++ views/admin/professions/top-level-information.njk
@@ -2,6 +2,7 @@
 
 {% from "govuk/components/input/macro.njk" import govukInput %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
+{% from "govuk/components/select/macro.njk" import govukSelect %}
 
 {% set bodyClasses = "rpr-internal__page" %}
 
@@ -29,6 +30,33 @@
           name: "name",
           value: name,
           errorMessage: errors.name | tError
+        })
+      }}
+
+      {{
+        govukSelect({
+          label: {
+            text: ("professions.form.label.topLevelInformation.regulatedAuthority" | t),
+            classes: "govuk-label--m",
+            isPageHeading: false
+          },
+          id: "regulatoryBody",
+          name: "regulatoryBody",
+          items: regulatedAuthoritiesSelectArgs,
+          errorMessage: errors.regulatoryBody | tError
+        })
+      }}
+
+      {{
+        govukSelect({
+          label: {
+            text: ("professions.form.label.topLevelInformation.additionalAuthority" | t),
+            classes: "govuk-label--m",
+            isPageHeading: false
+          },
+          id: "additionalRegulatoryBody",
+          name: "additionalRegulatoryBody",
+          items: additionalRegulatedAuthoritiesSelectArgs
         })
       }}
